Fix turning the world around (kind of).
Now the turning doesn't work correctly as it jumps "back" into the
center position when one clicks. This is because the projection.origin
method was removed before d3.v3 (I think) and I am not sure what the
equivalent would be. I thought it would be projection.rotate, but that
has it's own problems (seemingly "adding" rotation when used in place of
projection.center, which we currently use as a workaround).